refactor: :art: mini_type.h header added for typedef
mini_type.h header added for typedef
