+++ exec/envp_util.c
@@ -6,11 +6,11 @@
 /*   By: minjacho <minjacho@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/06 19:37:32 by minjacho          #+#    #+#             */
-/*   Updated: 2024/01/06 19:37:51 by minjacho         ###   ########.fr       */
+/*   Updated: 2024/01/06 20:35:01 by minjacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "mini_exec.h"
+#include "minishell.h"
 
 static char	*generate_env_str(t_dict *env_dict)
 {
