+++ exec/exec_errhandle.c
@@ -6,11 +6,11 @@
 /*   By: minjacho <minjacho@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/06 17:23:09 by minjacho          #+#    #+#             */
-/*   Updated: 2024/01/06 18:38:07 by minjacho         ###   ########.fr       */
+/*   Updated: 2024/01/06 20:35:00 by minjacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "mini_exec.h"
+#include "minishell.h"
 
 void	exit_with_str(char *str, int custom_errno)
 {
