+++ exec/exec_heredoc.c
@@ -6,11 +6,11 @@
 /*   By: minjacho <minjacho@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/06 19:04:20 by minjacho          #+#    #+#             */
-/*   Updated: 2024/01/06 19:12:06 by minjacho         ###   ########.fr       */
+/*   Updated: 2024/01/06 20:34:59 by minjacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "mini_exec.h"
+#include "minishell.h"
 
 void	heredoc_sub_preprocess(t_redir_node *redirect, int *cnt)
 {
